requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
cloudinary==1.44.1
